% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_weighting.R
\name{census_weighting}
\alias{census_weighting}
\title{Distance-weighting of spatial census variable}
\usage{
census_weighting(
  isochrones,
  tag = "tag",
  time = "time",
  census,
  b = 8,
  m = 0.5
)
}
\arguments{
\item{isochrones}{object of class \code{sf} containing road features,
derived from the \code{\link[DRIGLUCoSE]{isodistances}} function.}

\item{tag}{character or NA; string containing the column name, that indicates the unique tag column.}

\item{time}{character; string containing the column name, that indicates the time column.}

\item{b}{numeric; slope of distance decay function}

\item{m}{numeric; x for g(x) = 0.5}

\item{census_sf}{object of class \code{sf} containing a census shapefile.}
}
\value{
A \code{tibble} containing the spatially weighted census variable(s) of the
    intersects of the isochrone and Landsat objects.
}
\description{
To account for the diminishing effect of spatial variables (e.g. SES exposure) as
    distance increases, we fitted a logit function to weight each incremental isochrone. From
    a census shapefile, the spatially weighted areal mean of the isochrones are calculated.
}
\details{
For a detailed explanation of this method, see documentation on GitHub.
}
