% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlangen.R, R/erlangen_isodistances.R
\docType{data}
\name{Erlangen}
\alias{Erlangen}
\title{Sample Data for Analysis}
\format{
A \code{sf} with 2 column, which are:
\describe{
\item{tag}{id column}
\item{Speed}{Sex and age specific walking speed based on \href{https://ij-healthgeographics.biomedcentral.com/articles/10.1186/1476-072X-11-43}{Dewulf et al.}.}
}

A \code{sf} with 2 column, which are:
\describe{
\item{tag}{id column}
\item{time}{Isodistances}
}
}
\usage{
Erlangen

Erlangen
}
\description{
A simple \code{sf} of 2 points in Erlangen, Germany,
    that is for simple testing and running.

A simple \code{sf} of isodistances for 2 points in Erlangen, Germany,
    calculated using the \code{\link[DRIGLUCoSE]{isodistance}} function.
}
\keyword{datasets}
