% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrones.R
\name{isochrones}
\alias{isochrones}
\title{Calculate isochrones}
\usage{
isochrones(x, tag = "tag", buffer = 30, cores = 1)
}
\arguments{
\item{x}{object of class \code{sf} containing isodistances, derived from the \code{\link[DRIGLUCoSE]{isodistances}} function.}

\item{buffer}{numeric; buffer distance to be applied on the polyline features of the isodistance object.}

\item{cores}{the number of cores to use.}
}
\value{
An \code{sf MULTIPOLYGON} of isodistances is returned. The sf object contains two fields:
    tag (id of the original point feature) and time (range of the segments).
}
\description{
Calculates isochrones from \code{\link[DRIGLUCoSE]{isodistances}} by applying a off-road buffer to all polyline segments.
}
