% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodistances.R
\name{isodistances}
\alias{isodistances}
\title{Calculate isodistances}
\usage{
isodistances(
  x,
  road_network,
  speed = 78.5,
  tag = NA,
  isochrones_seq = c(5, 10, 15, 20),
  cores = 1
)
}
\arguments{
\item{x}{object of class \code{sf} of the origin point(s). Must contain one or more \code{POINT} features.}

\item{road_network}{object of class \code{sf} containing road features, derived from the \code{\link[DRIGLUCoSE]{osm_roads}} function.}

\item{speed}{numeric or character; either numeric value of speed (meters/minute) or string containing the column name, that indicates the speed.}

\item{tag}{character or NA; string containing the column name, that indicates the unique tag column.}

\item{isochrones_seq}{a numeric vector of isochrone values (in minutes).}

\item{cores}{the number of cores to use.}
}
\value{
An \code{sf MULTILINESTRING} of isodistances is returned. The sf object contains two fields:\cr
    tag (id of the original point feature) and time (range of the segments).
}
\description{
Based on a road network (\code{\link[DRIGLUCoSE]{osm_roads}}), this function calculates isodistances.
    An isodistance is a polyline of equal distance from a given point. All street segments belonging to a range will be summarized.
}
