% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS_L1C.R, R/LS_band_weighting.R
\name{LS_band_weighting}
\alias{LS_band_weighting}
\title{Distance-weighting of Landsat variable}
\usage{
LS_band_weighting(
  isochrones,
  tag = "tag",
  time = "time",
  landsat_list,
  band = "NDVI",
  b = 8,
  m = 0.5,
  cores,
  stats = list("sd", "median", list("percentile", 0.05), list("percentile", 0.95),
    "skew")
)

LS_band_weighting(
  isochrones,
  tag = "tag",
  time = "time",
  landsat_list,
  band = "NDVI",
  b = 8,
  m = 0.5,
  cores,
  stats = list("sd", "median", list("percentile", 0.05), list("percentile", 0.95),
    "skew")
)
}
\arguments{
\item{isochrones}{object of class \code{sf} containing road features,
derived from the \code{\link[DRIGLUCoSE]{isodistances}} function.}

\item{tag}{character or NA; string containing the column name, that indicates the unique tag column.}

\item{time}{character; string containing the column name, that indicates the time column.}

\item{landsat_list}{list of landsat products, derived from the \code{\link[DRIGLUCoSE]{LS_L1C}} function. See Details.}

\item{band}{character; spectral band or index from the landsat object,
to which distance-weighting should be applied.}

\item{b}{numeric; slope of distance decay function}

\item{m}{numeric; x for g(x) = 0.5}

\item{cores}{the number of cores to use.}

\item{stats}{The function to be applied. See Details}
}
\value{
A \code{tibble} containing the spatially weighted statistics of the
    intersects of the isochrone and Landsat objects.

A \code{tibble} containing the spatially weighted statistics of the
    intersects of the isochrone and Landsat objects.
}
\description{
To account for the diminishing effect of spatial variables (e.g. NDVI exposure) as
    distance increases, we fitted a logit function to weight each incremental isochrone.

To account for the diminishing effect of spatial variables (e.g. NDVI exposure) as
    distance increases, we fitted a logit function to weight each incremental isochrone.
}
\details{
All unique tags from the isochrones shapefile will be overlapped with the \code{landsat_list},
    and only the first overlapping landsat image will be used. This allows you to analyze multiple
    areas at once.\cr
    \cr
    Supported functions for the \code{stat} object:\cr
    "sum", "mean", "min", "max", "sd", "rms", "skew", "median" and "percentile". When using "percentile",
    provide a \code{list} as follows:\cr
    \code{list("percentile", upper_lim)}, where \code{upper_lim} is the upper limit of the percentile (e.g. 0.95).

All unique tags from the isochrones shapefile will be overlapped with the \code{landsat_list},
    and only the first overlapping landsat image will be used. This allows you to analyze multiple
    areas at once.\cr
    \cr
    Supported functions for the \code{stat} object:\cr
    "sum", "mean", "min", "max", "sd", "rms", "skew", "median" and "percentile". When using "percentile",
    provide a \code{list} as follows:\cr
    \code{list("percentile", upper_lim)}, where \code{upper_lim} is the upper limit of the percentile (e.g. 0.95).
}
