% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS_L1C.R
\name{LS_L1C}
\alias{LS_L1C}
\title{Correct Landsat L1C products}
\usage{
LS_L1C(
  l1c_path = NULL,
  out_dir = NULL,
  proc_dir = NULL,
  sf_mask = NA,
  bad_pixel = TRUE,
  dark_pixel = TRUE,
  indices = c("MNDWI", "NDVI"),
  cores = 1L,
  maxmemory = 1e+08
)
}
\arguments{
\item{l1c_path}{Full/relative path of L1C product. This can be both the original .tar.gz file or the unzipped file.}

\item{out_dir}{Directory where output L2A products will be placed.}

\item{proc_dir}{(optional) Directory where processing is applied. If NULL (default),
processing is done in \code{l1c_path}-directory and output L2A product is then moved to \code{out_dir}.}

\item{sf_mask}{\code{sf} object; CRS from this sf object will be used to project all raster layers.
This process is computationally expensive, therefore parallel computation is applied. The raster will also be cropped and
masked to the bounding box of the \code{sf_mask} object.}

\item{bad_pixel}{Remove pixels with values of <0.}

\item{dark_pixel}{The Dark Object Subtraction method assumes that the darkest parts of an image
(water, artificial structures) should be black, if not for the effects of atmospheric scatter.
The lowest value of each band will therefore be subtracted.}

\item{indices}{(optional) Character string or vector of indices to calculate.\cr
\code{MNDWI}: Modified Normalized Difference Water Index: \href{https://doi.org/10.1080/01431160600589179}{Xu H.}\cr
\code{NDVI}: Normalized Difference Vegetation Index:
\href{https://gisgeography.com/ndvi-normalized-difference-vegetation-index/}{GIS Geography}}

\item{cores}{The number of cores to use; only necessary if \code{sf_mask} is provided.}

\item{maxmemory}{numeric; Maximum number of bytes to read into memory. If a process is expected to require more
than this value, \code{\link[raster]{canProcessInMemory}} will return \code{FALSE.}}
}
\value{
L2A.grd file in \code{out_dir}.
}
\description{
The function corrects Landsat 1-8 L1C images to L2A.
    For the digital number (DN) to top of atmosphere (TOA) reflectance I used the metadata
    from the \code{MTL.txt} with the following equation:\cr
    TOA_RAD = rawDN * RADIANCE_MULT_BAND + RADIANCE_ADD_BAND /sin(SUN_ELEVATION * 180)\cr
    \cr
    At this point, only the following bands will be saved:\cr
    Red, Green, Blue, NIR, SWIR1, SWIR2
}
