% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_roads.R
\name{osm_roads}
\alias{osm_roads}
\title{Download and clean road features}
\usage{
osm_roads(
  x,
  dist,
  speed,
  cores = 1L,
  remove_features = c("motorway", "motorway_link", "trunk", "trunk_link", "raceway"),
  split_segments = TRUE
)
}
\arguments{
\item{x}{object of class \code{sf}.}

\item{dist}{numeric; maximum number of minutes of the isochrones.}

\item{speed}{numeric or character; either numeric value of speed or string containing the column name, that indicates the walking speed.}

\item{cores}{the number of cores to use.}

\item{remove_features}{character vector containing feature keys that should be excluded from the analysis (e.g. motorway)}

\item{split_segments}{logical; should topology cleaning with \code{\link[nngeo]{st_segments}} be used.}
}
\value{
object of class \code{sf} containing road features.
}
\description{
Download road features using the OSM overpass API \code{\link[osmdata]{opq}}.
    Simple topology cleaning is applied using \code{\link[nngeo]{st_segments}}.
}
\examples{
\dontrun{
data(Erlangen)
osm_roads(Erlangen, dist = 20, speed = "Speed", cores = 4)
}
}
