% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_parallel.R
\name{rbind_parallel}
\alias{rbind_parallel}
\title{Parallel rbind}
\usage{
rbind_parallel(X, cores = 1L)
}
\arguments{
\item{X}{a \code{list} object.}

\item{cores}{the number of cores to use.}
}
\description{
Helper function to allow parallel \code{\link{rbind}}
}
