% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlangen.R
\docType{data}
\name{Erlangen}
\alias{Erlangen}
\title{Sample Data for Analysis}
\format{
A \code{sf} with 2 column, which are:
\describe{
\item{id}{id}
\item{Speed}{Sex and age specific walking speed based on \href{https://ij-healthgeographics.biomedcentral.com/articles/10.1186/1476-072X-11-43}{Dewulf et al.}.}
}
}
\usage{
Erlangen
}
\description{
A simple \code{sf} of 2 points in Erlangen, Germany,
    that is for simple testing and running.
}
\keyword{datasets}
